Labfiles/01-analyze-text/Python/text-analysis/requirements.txt
python-dotenv
azure-ai-textanalytics
azure-identity
azure-core